requirements.txt
gunicorn
django-heroku
postgres
psycopg2
pytz
whitenoise
social-auth-app-django
Pillow
